+++ test/src/forum/service/SectionService.java
@@ -9,15 +9,15 @@ public interface SectionService {
 	public void saveSection(Section section);
 	
 	public void deleteSection(Section section);
-	public void deleteSectionById(Integer sectionId);
+	public void deleteSectionById(long sectionId);
 	
 //	public List<Section> findSectionList(Section s_section,PageBean pageBean);
 	public List<Section> findSectionList(Section s_section,int pageSize,int pageNo);
 	
-	List<Section> findSectionListByZoneId(int zoneId,int pageSize,int pageNo);
+	List<Section> findSectionListByZoneId(long zoneId,int pageSize,int pageNo);
 	
 	public Long getSectionCount(Section s_section);
 	
-	public Section findSectionById(int sectionId);
+	public Section findSectionById(long sectionId);
 	
 }
