+++ test/src/forum/service/ZoneService.java
@@ -10,11 +10,11 @@ public interface ZoneService {
 	
 	public void deleteZone(Zone zone);
 	
-	public void deleteZoneById(Integer id);
+	public void deleteZoneById(long id);
 	
 	public List<Zone> findZoneList(Zone s_zone,int pageSize,int pageNo);
 	
 	public Long getZoneCount(Zone s_zone);
 	
-	public Zone findZoneById(int zoneId);
+	public Zone findZoneById(long zoneId);
 }
